requirements.txt
google-generativeai
PyPDF2